move wasm specific code to .wasm.cs
revert merge error
